+++ vehicle/src/Vehicle/Backend/Solver/QueryCompilation/InputBoundsCheck.hs
@@ -9,9 +9,11 @@ import Control.Monad.Reader (MonadReader (..), ReaderT (..))
 import Data.Coerce (coerce)
 import Data.Either (partitionEithers)
 import Data.Foldable (foldlM)
+import Data.List (singleton)
 import Data.List.NonEmpty (NonEmpty (..))
 import Data.Map (Map)
 import Data.Map qualified as Map
+import Data.Maybe (fromMaybe)
 import Data.Set qualified as Set
 import Vehicle.Backend.Solver.QueryCompilation.Core (MonadQueryCompilation)
 import Vehicle.Backend.Solver.UserVariableElimination.Core
@@ -23,10 +25,11 @@ import Vehicle.Compile.Resource (NetworkName)
 import Vehicle.Data.Bound
 import Vehicle.Data.Bound.FourierMotzkinElimination (fourierMotzkinTensorBoundsElimination)
 import Vehicle.Data.Code.BooleanExpr
+import Vehicle.Data.Code.LinearExpr (ConstantLike)
 import Vehicle.Data.Code.Value (boundVariablesIn)
 import Vehicle.Data.MaybeTrivial (MonadMaybeTrivial (..))
 import Vehicle.Data.Tensor (HasShape (..), RatTensor, TensorShape)
-import Vehicle.Data.Tensor.Traversal (toPartialShape)
+import Vehicle.Data.Tensor.Traversal (MonadTraverseTensor, PartiallyKnownTensorShape, toPartialShape)
 import Vehicle.Data.Variable.Bound.Context.Name
 import Vehicle.Data.Variable.Bound.Context.Tensor.Class (MonadReadableTensorBoundContext, getCompleteNamedCtx, lookupParentTensorVariable)
 import Vehicle.Data.Variable.Bound.Level
@@ -43,7 +46,7 @@ findInputVariableBounds ::
   (MonadQueryCompilation m, MonadMaybeTrivial m) =>
   NetworkApplications ->
   ConjunctAll LinearAssertion ->
-  m (BoundedAssertions NetworkInputTensorVariable SliceVariable RatTensor)
+  m (BoundedAssertions SliceVariable RatTensor)
 findInputVariableBounds metaNetworkApps constraints = do
   logCompilerSection2 MaxDetail "network variable bounds checks" $
     runMonadBoundsT metaNetworkApps $ do
@@ -54,7 +57,12 @@ findInputVariableBounds metaNetworkApps constraints = do
         let Partial bounds assertions = boundedAssertions
         assertionsDoc <- prettyFriendlyInCtx assertions
         boundsDoc <- prettyBounds bounds
-        return $ "Found bounds:" <> lineIndent boundsDoc <> line <> "remaining assertions:" <> lineIndent assertionsDoc
+        return $
+          "Found bounds:"
+            <> lineIndent boundsDoc
+            <> line
+            <> "remaining assertions:"
+            <> lineIndent assertionsDoc
 
       -- Check that all bounds present
       checkAllBoundsPresent boundedAssertions
@@ -180,7 +188,7 @@ lookupCorrespondingInputVar var = do
 checkAllBoundsPresent ::
   (MonadBounds m, MonadMaybeTrivial m) =>
   PartiallyBoundedAssertions ->
-  m (BoundedAssertions NetworkInputTensorVariable SliceVariable RatTensor)
+  m (BoundedAssertions SliceVariable RatTensor)
 checkAllBoundsPresent (Partial allPartialbounds assertions) = do
   (PropertyMetaData {..}, _, inputVariableMapping) <- ask
 
@@ -192,8 +200,8 @@ checkAllBoundsPresent (Partial allPartialbounds assertions) = do
         let partialShape = toPartialShape varShape Nothing
         missingIndicesOrFlattenedBounds <- fourierMotzkinTensorBoundsElimination partialShape partialBounds
         case missingIndicesOrFlattenedBounds of
-          Right bounds -> return $ Right (BoundedValue var bounds)
           Left missingIndices -> errorCase missingIndices
+          Right _bounds -> Right <$> lowerTensorBounds partialBounds varShape
 
   let (missingBounds, boundedVariables) = partitionEithers errorsAndFinalBounds
   case missingBounds of
@@ -204,13 +212,13 @@ checkAllBoundsPresent (Partial allPartialbounds assertions) = do
 
   -- If Marabou, then warn if all inputs are constant.
   -- See https://github.com/NeuralNetworkVerification/Marabou/issues/670
-  let domains = fmap valueBounds boundedVariables
+  let domains = fmap (fmap valueBounds) boundedVariables
   let formatID = queryFormatID queryFormat
-  when (queryFormatID queryFormat == MarabouQueries && all isEquality domains) $
+  when (queryFormatID queryFormat == MarabouQueries && all (all isEquality) domains) $
     logWarning $
       AllConstantNetworkInputVars formatID propertyAddress
 
-  let boundsUnsatisfiable = not (all isSatisfiable domains)
+  let boundsUnsatisfiable = not (all (all isSatisfiable) domains)
   if boundsUnsatisfiable
     then trivial False
     else case assertions of
@@ -218,7 +226,7 @@ checkAllBoundsPresent (Partial allPartialbounds assertions) = do
       a : as ->
         return $
           BoundedAssertions
-            { variableBounds = boundedVariables,
+            { variableBounds = concat boundedVariables,
               assertions = ConjunctAll (a :| as)
             }
 
@@ -227,3 +235,95 @@ findUnboundedVariables appInfo =
   -- TODO we actually need to do this recursively on any network variables that
   -- live in this set.
   Set.toList $ boundVariablesIn $ inputValue appInfo
+
+-- | Takes in bounds over a tensor and tries to compute a single lower and upper bound for the
+-- whole tensor. If it fails then it returns a list of the indices of the tensor which
+-- are unbounded.
+--
+-- NOTE: this function is currently unsound as at the moment it discards the strictness information.
+-- See https://github.com/vehicle-lang/vehicle/issues/74
+lowerTensorBounds ::
+  (MonadLogger m, MonadReadableNameContext m) =>
+  TensorBounds RatTensor ->
+  TensorShape ->
+  m [BoundedValue NetworkIOElementVariable (Domain Rational)]
+lowerTensorBounds TensorBounds {..} =
+  lowerNestedSliceBounds (SliceBounds mempty mempty) (Just tensorSliceBounds)
+
+lowerNestedSliceBounds ::
+  forall m.
+  (MonadLogger m, MonadReadableNameContext m) =>
+  SliceBounds RatTensor ->
+  Maybe (NestedSliceBounds RatTensor) ->
+  TensorShape ->
+  m [BoundedValue NetworkIOElementVariable (Domain Rational)]
+lowerNestedSliceBounds inheritedBounds maybeNestedSliceBounds shape = do
+  let currentLowerBounds = lowerBounds inheritedBounds <> maybe mempty (lowerBounds . sliceBounds) maybeNestedSliceBounds
+  let currentUpperBounds = upperBounds inheritedBounds <> maybe mempty (upperBounds . sliceBounds) maybeNestedSliceBounds
+  let maybeLowerBound = andBoundList currentLowerBounds
+  let maybeUpperBound = andBoundList currentUpperBounds
+
+  case shape of
+    [] -> do
+      let allBounds = _
+      return [BoundedValue _ (Domain _ _)]
+    d : ds -> do
+      let maybeChildBounds = childSliceBounds =<< maybeNestedSliceBounds
+      case maybeChildBounds of
+        Nothing -> return $ concat <$> traverse _ _
+        Just xs -> _
+
+{-
+case maybeChildBounds of
+  Nothing -> case andBoundList (getBound sliceBounds) of
+    Just bounds -> return $ Right bounds
+    Nothing -> do
+      indices <- currentIndices
+      return $ Left [indices]
+  Just childTensorBounds -> do
+    childErrorOrBounds <- traverseTensorRows go childTensorBounds
+    let (missingChildIndices, childBounds) = partitionEithers childErrorOrBounds
+    case andBoundList (getBound sliceBounds) of
+      Nothing -> case missingChildIndices of
+        i : is -> return $ Left $ concatNonEmpty $ i :| is
+        [] -> return $ Right $ stackBounds childBounds
+      Just bound -> do
+        case missingChildIndices of
+          _ : _ -> do
+            let boundElements = unstackBounds bound
+            let combineBound boundElement = either (const boundElement) (andBound boundElement)
+            let combinedElements = zipWith combineBound boundElements childErrorOrBounds
+            return $ Right $ stackBounds combinedElements
+          [] -> do
+            let childBound = stackBounds childBounds
+            return $ Right $ andBound bound childBound
+-}
+
+{-
+
+reduceDomain ::
+  forall m.
+  (MonadQueryCompilation m) =>
+  BoundedValue NetworkInputTensorVariable (Domain RatTensor) ->
+  m [BoundedValue NetworkIOElementVariable (Domain Rational)]
+reduceDomain (BoundedValue inputTensorVar bounds) = go (toSliceVar inputTensorVar, bounds)
+  where
+    go ::
+      (SliceVariable, Domain RatTensor) ->
+      m [BoundedValue NetworkIOElementVariable (Domain Rational)]
+    go (var, Domain lowerBound upperBound) =
+      case shapeOf lowerBound of
+        [] -> do
+          let elementVar = coerce var
+          let lowerValue = fmap extractRationalConstant lowerBound
+          let upperValue = fmap extractRationalConstant upperBound
+          let domain = Domain lowerValue upperValue
+          return [BoundedValue elementVar domain]
+        _ : _ -> do
+          childVars <- lookupChildVariablesCertain var
+          let lowerBounds = unstackBounds lowerBound
+          let upperBounds = unstackBounds upperBound
+          let domains = zipWith Domain lowerBounds upperBounds
+          concat <$> zipWithM (curry go) childVars domains
+
+-}